--- lib/rails_extensions/active_record/base.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-require "active_record"
-
-module ActiveRecord
-  class Base
-    class_attribute :store_base_sti_class
-    self.store_base_sti_class = true
-  end
-end
